move code around to make app.tsx smaller
