+++ src/app.tsx
@@ -1,59 +1,16 @@
-import * as React from "react"
-import { range, rangeFromTo } from "lang/ranges";
-import { removeFromArray, shuffle, unique } from "lang/arrays";
-import { noop, NoState } from "lang/react";
-import StackPopped from "stacks/stackPopped";
-import BringEventToConclusion from "events/bringEventToConclusion";
-import PartyBattleGoals from "battlegoals/partyBattleGoals";
-import AddCards, { AddCardsProps } from "cards/addCards";
-import "./style.css";
-
-class RandomSideScenarioProps {
-  readonly kind: string = 'random-side-scenario-props';
-  readonly notUsedByAnyOne: boolean = false; // if you delete this, typescript becomes un-happy
-}
-
-class RandomItemDesignProps {
-  readonly kind: string = 'random-item-design-props';
-  readonly url: string = "https://lh3.googleusercontent.com/u/0/d/1ubPFoTJ1_Ly-Eqn_yMvNhG0Dwq_tuw5c=s4800-k-iv1";
-  readonly width: number = 409;
-  readonly height: number = 636;
-  readonly offset: number = 71;
-  readonly cols: number = 10;
-  readonly n: number = 25;
-}
-
-class ItemUrl {
-  constructor(readonly url: string,
-              readonly numberInPicture: number,
-              readonly n: number) {
-  }
-}
-
-class PersonalGoalProps {
-  readonly kind: string = 'persona-goal-props';
-  readonly offset: number = 510;
-  readonly n: number = 24;
-  readonly divWidth: string = "605px";
-}
-
-class ItemProps {
-  readonly kind: string = 'item-props';
-  readonly url: string;
-  readonly width: number = 292;
-  readonly height: number = 456;
-  readonly cols: number = 10;
-  readonly n: number;
-
-  constructor(private item: ItemUrl, readonly offset: number) {
-    this.item = item;
-    this.url = item.url;
-    this.offset = offset;
-    this.n = item.n;
-  }
-}
-
-type CardRenderProps = RandomItemDesignProps | ItemProps | PersonalGoalProps | RandomSideScenarioProps;
+import * as React from 'react'
+import { range, rangeFromTo } from 'lang/ranges';
+import { removeFromArray, shuffle, unique } from 'lang/arrays';
+import { noop, NoState } from 'lang/react';
+import StackPopped from 'stacks/stackPopped';
+import BringEventToConclusion from 'events/bringEventToConclusion';
+import PartyBattleGoals from 'battlegoals/partyBattleGoals';
+import AddCards, { AddCardsProps } from 'cards/addCards';
+import { Draw } from 'cards/draw';
+import { CardStack } from 'cards/cards';
+import { OpenDialog } from 'app/app';
+import { RandomItemDesignProps, PersonalGoalProps, ItemUrl, ItemProps, RandomSideScenarioProps } from "cards/CardRenderProps";
+import { cardToDiv } from "cards/cards";
 
 const randomItemDesigns = new RandomItemDesignProps();
 const personalGoals = new PersonalGoalProps();
@@ -121,54 +78,6 @@ function itemToDiv(itemId: number) {
   return cardToDiv(itemId, new ItemProps(item, itemId - item.numberInPicture));
 }
 
-function isPersonalGoalProps(arg: CardRenderProps): arg is PersonalGoalProps {
-  return arg.kind === 'persona-goal-props';
-}
-
-function isRandomSideScenarioProps(arg: CardRenderProps): arg is RandomSideScenarioProps {
-  return arg.kind === 'random-side-scenario-props';
-}
-
-function cardToDiv(cardId: number, props: CardRenderProps) {
-  if (isPersonalGoalProps(props)) {
-    return <PersonalGoalCard cardId={cardId}/>
-  } else if (isRandomSideScenarioProps(props)) {
-    return null
-  } else {
-    let n = cardId - props.offset;
-    let row = Math.floor(n / props.cols);
-    let col = n % props.cols;
-    let style = {
-      borderRadius: '15px',
-      background: "url(" + props.url + ") no-repeat scroll top -" + (row * props.height) + "px left -" + (col * props.width) + "px",
-      width: (props.width - 14) + "px",
-      maxWidth: (props.width - 14) + "px",
-      height: (props.height - 3) + "px",
-      marginLeft: "10px",
-      color: "white",
-      padding: "0 0 3px 14px",
-      display: "inline-block",
-    };
-    return <div key={cardId} style={style}>{cardId}</div>;
-  }
-}
-
-interface PersonalGoalCardProps {
-  cardId: number;
-}
-
-class PersonalGoalCard extends React.Component<PersonalGoalCardProps, NoState> {
-  render() {
-    const style = {
-      display: "inline-block",
-      width: "605px",
-      borderRadius: '20px'
-    } as React.CSSProperties;
-    const cardId = this.props.cardId;
-    return <img key={cardId} style={style} src={"https://raw.githubusercontent.com/any2cards/gloomhaven/master/images/personal-goals/pg-" + cardId + ".png"} alt='personal goal'/>
-  }
-}
-
 interface RandomItemDesignsProps {
   list: Array<number>
 }
@@ -185,99 +94,6 @@ class RandomItemDesigns extends React.Component<RandomItemDesignsProps, NoState>
         </div>
       ]
     }
-  }
-}
-
-interface RandomCardProps {
-  drawnCards: Array<number>;
-  name: string;
-  drawn: (name: string, cards: CardStack, cardNo: number) => void;
-  cards: CardStack;
-  cardProps: CardRenderProps;
-}
-
-class RandomCard extends React.Component<RandomCardProps, NoState> {
-  clicked(cardNo: number) {
-    this.props.drawn(this.props.name, this.props.cards, cardNo);
-  }
-
-  render() {
-    return <React.Fragment>
-      <h2 key="h2">Drawn {this.props.name}: {this.props.drawnCards.join(" ")}</h2>
-      <div key="button-div">
-        {this.props.drawnCards.map(cardNumber => {
-          const styles = {
-            display: "inline-block"
-          } as React.CSSProperties;
-          if (isPersonalGoalProps(this.props.cardProps)) {
-            styles.width = this.props.cardProps.divWidth
-          }
-          return <div key={"span-" + cardNumber} style={(styles)}>
-            <div key={"button-div-" + cardNumber}>
-              <button key={"button-" + cardNumber} type="button" onClick={() => this.clicked(cardNumber)}>Accept {cardNumber}</button>
-            </div>
-            {cardToDiv(cardNumber, this.props.cardProps)}
-          </div>
-        })}
-      </div>
-    </React.Fragment>
-  }
-}
-
-interface OpenDialog {
-  setDialog: (component: JSX.Element) => void;
-}
-
-interface CardStackEvent {
-  action: string,
-  card?: number,
-  cards?: Array<number>,
-  event?: number,
-}
-
-export interface CardStack {
-  list: Array<number>,
-  stack: Array<number>,
-  history: Array<CardStackEvent>
-}
-
-interface DrawProps extends OpenDialog, DrawnCallback {
-  cards: CardStack;
-  n: number;
-  name: string;
-  cardProps: CardRenderProps;
-}
-
-// Draw draws a _random_ card from the deck
-class Draw extends React.Component<DrawProps, NoState> {
-  constructor(props: DrawProps) {
-    super(props);
-    this.clicked = this.clicked.bind(this);
-  }
-
-  clicked() {
-    let drawnCards = [];
-    let next;
-    for (let i = 0; (i < this.props.n) && (drawnCards.length < this.props.cards.stack.length); i++) {
-      do {
-        next = this.props.cards.stack[Math.floor(Math.random() * this.props.cards.stack.length)];
-      } while (drawnCards.indexOf(next) > -1);
-      drawnCards.push(next);
-    }
-    drawnCards.sort();
-    this.props.setDialog(
-      <RandomCard
-        name={this.props.name}
-        drawnCards={drawnCards}
-        cards={this.props.cards}
-        drawn={this.props.drawn}
-        cardProps={this.props.cardProps}
-      />
-    );
-  }
-
-  render() {
-    return <button type="button" onClick={this.clicked} disabled={this.props.cards.stack.length === 0}>{"Draw " + this.props.name}</button>
   }
 }
 
@@ -482,10 +298,6 @@ function itemsAboveProsperity(title: String, items: Array<number>, prosperity: n
   </div>;
 }
 
-interface DrawnCallback {
-  drawn: (name: String, cards: CardStack, cardNo: number) => void;
-}
-
 export interface CardStacks {
   cityEvents: CardStack;
   roadEvents: CardStack;
@@ -688,7 +500,7 @@ export class App extends React.Component<AppProps, AppState> {
       }
 
       const notAlreadyContainedCards = unique(cardIdsToAdd).filter((it) => !stack.stack.includes(it));
-      if(notAlreadyContainedCards.length === 0 ) {
+      if (notAlreadyContainedCards.length === 0) {
         return;
       }
