+++ src/app/app.ts
@@ -1,0 +1,3 @@
+export interface OpenDialog {
+  setDialog: (component: JSX.Element) => void;
+}
